feat: ability to specify extra_targets for skeletonization
feat: specify root for core skeletonization routine

This allows an end user to specify e.g. the location of
synapses in a neuron to ensure we trace to them. Note that
unlike border_targets, extra_targets do not have an effect
on what the root node will be. This is because they are totally
uncontrolled and can be located off of the shape.