feat: specify extra_targets_before and after (#33)
* feat: ability to specify extra_targets for skeletonization

feat: specify root for core skeletonization routine

This allows an end user to specify e.g. the location of
synapses in a neuron to ensure we trace to them. Note that
unlike border_targets, extra_targets do not have an effect
on what the root node will be. This is because they are totally
uncontrolled and can be located off of the shape.

* fix: ensure soma_radius is always set

* redesign: change extra_targets to be list of points

Originally { label: [ pts... ] }, but this doesn't account for
connected components nor the possibility of not having
the label mapping to begin with.

* fix: extra_targets need to be translated to ROI

* feat: introduce extra_targets_before and extra_targets_after

* test: super basic sanity test for before and after

* docs: describe extra_targets_after and extra_targets_before