feat: facility for converting synapse centroids into targets (#37)
* feat: facility for converting synapse centroids into targets

Often synapses are specified as (xyz_centroid, prelabel, postlabel),
however this point is usually located within the cleft between two
neurons. This is not a useful point for skeletonization. However,
as a heuristic, we can pick the closest point to the synapse centroid
for the pre and post labels.

This doesn't work if the centroid gets so wacky that it's far away from
the in-between space of the cells, but this should be unlikely to happen
with high quality data.

* fix: deal with labels not inside the volume

* refactor+perf: use new fill-voids package

A faster and lower memory implementation of kimimaro.skeletontricks.binary_fill_holes

* fix: correctly assign swc_labels to adjusted synapse targets